Add basic languages for concrete and gradual types
As in the AGT paper, the language is first defined in a concrete way,
and then in an abstract way. The abstracted form can be applied to the
identity unit functor to get back to the concrete form. Lowering the
relations back from their power-lifted form is not yet implemented, and
may require a change to the existing abstraction.