Split out functor definitions, add All predicate
The UnitFunctor definitions are now in their own module. The abstracted
language implementation is now called ATFL. ATFL is applied to the
Identity unit functor under the equality relation to produce STFL. STFL
is then used to define the abstraction function γ for ATFL applied to
the Gradual unit functor. The abstraction function produces consistent
equality, which is used to generate the GTFL. As the generation of the
concrete languages do not refer to the constructors of ATFL's Type, the
next step is to move the generation of Identity and Gradual languages
into an abstracted form, redefine ATFL as an implementation of their
corresponding dependency, and then further refine ATFL into STFL and
GTFL as applications of the abstracted generation.