Replace All with a generic map operation
The All proposition required in ATFL to define the concretisation
function γ has been replaced with the more generic map function that,
given a proof that a RecType is indexed by a functor F and is a natural
transformation from F to another arbitrary functor G (but with both
specialised to the contained type RecType G), can transform any RecType
indexed by F into one indexed by G. This is good because it both
justifies the use of functors (as up until now only the unit part was
being used), and allows the continued definition of γ. The All type is
replaced by embedding proofs about γ in a type, and then decomposing
those proofs with map into their component gradual and concrete types.

The constant functor is now included in the Category.UnitFunctor
module (though it is not a UnitFunctor), as this is the functor F in the
application to map above. The embedded proofs ignore the recursive
component of RecType to directly embed the relevant proof types.