Split functor modules, make language abstract
The Category.UnitFunctor module has been split into an accompanying
Category.Enfofunctor module, which defines the type of a Haskell-like
endofunctor and some examples, which is then embedded in an instance of
UnitFunctor.

The example language in Type.Gradual is now defined entirely abstractly
as a RecNatTrans, which supplies a very specific instance of a natural
transformation map in order to implement γ in a way that is completely
divorced from the language. Opening the Gradual inner module with the
now Abstractly Typed Functional Language's type definition automatically
gives the appropriate instance of γ. Examples of consistency and terms
are now provided.

The power predicate's concrete values can usually be inferred, so these
arguments to the constructor are now implicit.