Correct some predicates, split terms from typing
The 'partial function' predicates weren't correct for allowing the left
side of a function call to have an unknown type, because it performed an
implicit equality relation. These now use the local consistency instead,
which automatically lifts these definitions (but without the appropriate
precision).

Because a lot of typical implicit equalities in Term declarations need
to be moved out into their own enforcement of local consistency, typing
is now separated from the Term type in the example language, so
ill-typed terms can be created but not typed.