Abstract the implementation of γ to any functor
The implementation of γ was implied to be implemented as a value of the
Gradual type, as it had the same constructors. Now the inner Gradual
module is the Abstract module, implemented for any given functor F, with
γ implemented as a lifting of a given instance of F applied to 'Tγ', now
with the requirement that the concrete component is equal to a given
concrete type. The standard library implementation of Maybe is pulled in
and implemented as both an endofunctor and unit functor, and used to
implement GTFL.