Simplify language modules, use instance parameter
The Under module between the outer ATFL module and the inner Language
module has been removed, allowing the Language record to be opened
directly. It can either be opened without binding the necessary unit
functor (in which case the language remains ATFL) or bound (STFL, GTFL,
DTFL etc). The example has been updated to bind the gradual functor (now
called Maybe) and open the language with consistent-lifted equality.

The functor parameter to the RecNatTrans map function is now an instance
parameter, so it may be omitted if a relevant functor instance is
around (though in the one application of it in γ it cannot be inferred).
The functor declarations in both Endofunctor and UnitFunctor modules
have been updated to instance declarations.