Move the definition of lifted equality into ATFL
Each concrete definition using the ATFL no longer has to define lifted
equality, it is now defined directly in the ATFL. The whole module is
defined in terms of some unit functor, though the definitions for Type
and map are still abstracted to any functor. The lifted equality is now
guaranteed to perform as a consistent equality of some kind, depending
only on the value of the functor.